Added new class Teacher and changed Student class to AncientStudent class
