+++ src/com/rafi/app/Main.java
@@ -4,10 +4,11 @@ public class Main {
 
 	public static void main(String[] args) {
 
-		Student student = new Student(123, "Moga");
-		
-		System.out.println(student);
-		
+		Teacher teacher = new Teacher();
+		teacher.setName("Wise teacher");
+		System.out.println("I am a " + teacher.getName());
+			
+		AncientStudent student = new AncientStudent("Moga");
 		student.hanashi();
 
 	}
