--- src/com/rafi/app/Student.java
@@ -1,45 +1,0 @@
-package com.rafi.app;
-
-public class Student {
-
-	private int roll;
-	private String name;
-
-	public Student() {
-	}
-
-	public Student(int roll, String name) {
-		super();
-		this.roll = roll;
-		this.name = name;
-	}
-
-	public int getRoll() {
-		return roll;
-	}
-
-	public void setRoll(int roll) {
-		this.roll = roll;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-	
-	public void hanashi(){
-		
-		System.out.println("Ami "+ name + ".amar Roll "+ roll);
-	}
-
-	@Override
-	public String toString() {
-		return "Student [roll=" + roll + ", name=" + name + "]";
-	}
-	
-	
-
-}
