+++ src/com/rafi/app/Teacher.java
@@ -1,0 +1,13 @@
+package com.rafi.app;
+
+public class Teacher {
+	private String name;
+	
+	public String getName(){
+		return name;
+	}
+	
+	public void setName(String name){
+		this.name = name;
+	}
+}
